refactor: main의 코드를 관심사에 따라 분리
- movieStore로 상태 관리
- movieService로 API 호출
- movieUI로 DOM 업데이트
